AdversarialML/MJA/requirements.txt
PyYAML==6.02
python-dotenv==1.1.0
transformers==4.43.1
huggingface_hub==0.32.2
pandas==2.2.2
torch==2.3.1
scipy==1.11.4
scikit-learn==1.3.2